% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{ask_question}
\alias{ask_question}
\title{Prompt User for Class Conversion Options}
\usage{
ask_question(var)
}
\arguments{
\item{var}{The name of the variable for which the user is prompted to select a class conversion option.}
}
\value{
None. This function is used for prompting the user and does not return a value.
}
\description{
This function prompts the user with options to convert a variable to a specific class (e.g., factor, integer, numeric, character, or logical).
The function provides a list of class conversion options for the specified variable and includes an option to cancel the operation.
}
\examples{
ask_question("variable_name")
}
