% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{ask_question_wait_response_class}
\alias{ask_question_wait_response_class}
\title{Ask Question and Wait for Class Response}
\usage{
ask_question_wait_response_class(var)
}
\arguments{
\item{question}{The question to ask the user.}
}
\value{
The user's decision.
}
\description{
Prompts the user with a question and waits for a response related to class decisions.
}
