% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{ask_question_wait_response_levels}
\alias{ask_question_wait_response_levels}
\title{Ask Question and Wait for Response on Factor Levels}
\usage{
ask_question_wait_response_levels(level_conflicts)
}
\arguments{
\item{level_conflicts}{A list of variables with factor level conflicts.}
}
\value{
The user's decision.
}
\description{
Prompts the user with options for resolving factor level conflicts and waits for a response.
}
