% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{check_response_class}
\alias{check_response_class}
\title{Check User Response for Class Decision}
\usage{
check_response_class(answer, var)
}
\arguments{
\item{answer}{The user's input.}

\item{var}{The variable name.}
}
\value{
The user's decision or a recursive prompt for input.
}
\description{
Checks the user's input to ensure it is valid for class decisions.
}
