% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{check_response_levels}
\alias{check_response_levels}
\title{Check User Response for Factor Levels}
\usage{
check_response_levels(answer, level_conflicts)
}
\arguments{
\item{answer}{The user's input.}

\item{level_conflicts}{A list of variables with factor level conflicts.}
}
\value{
The user's decision.
}
\description{
Checks the user's input to ensure it is valid for resolving factor level conflicts.
}
