% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.add_missing_cols_to_df}
\alias{.add_missing_cols_to_df}
\title{Add Missing Columns to DataFrame}
\usage{
.add_missing_cols_to_df(df.name, unique_cols, newobj, datasources)
}
\arguments{
\item{df.name}{The name of the DataFrame.}

\item{unique_cols}{A vector of unique column names.}

\item{newobj}{The name of the new DataFrame.}

\item{datasources}{Data sources from which to aggregate data.}
}
\value{
None. Updates the DataFrame with added columns.
}
\description{
Adds any missing columns to the DataFrame to ensure all columns are present across data sources.
}
