% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.fix_classes}
\alias{.fix_classes}
\title{Fix Variable Classes}
\usage{
.fix_classes(df.name, different_classes, class_decisions, newobj, datasources)
}
\arguments{
\item{df.name}{The name of the DataFrame.}

\item{different_classes}{A list of variables with class conflicts.}

\item{class_decisions}{The decisions made by the user.}

\item{newobj}{The name of the new DataFrame.}

\item{datasources}{Data sources from which to aggregate data.}
}
\value{
None. Updates the DataFrame with consistent variable classes.
}
\description{
Applies the user's class decisions to fix the classes of variables across different data sources.
}
