% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.get_added_cols}
\alias{.get_added_cols}
\title{Get Added Columns}
\usage{
.get_added_cols(old_names, new_names)
}
\arguments{
\item{old_names}{A list of old column names.}

\item{new_names}{A list of new column names.}
}
\value{
A list of added column names.
}
\description{
Compares the old and new column names and identifies newly added columns.
}
