% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.get_factor_levels}
\alias{.get_factor_levels}
\title{Get Factor Levels from Data Sources}
\usage{
.get_factor_levels(factor_vars, newobj, datasources)
}
\arguments{
\item{factor_vars}{A vector of factor variables.}

\item{newobj}{The name of the new DataFrame.}

\item{datasources}{Data sources from which to aggregate data.}
}
\value{
A list of factor levels.
}
\description{
Retrieves the levels of factor variables from different data sources.
}
