% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.get_unique_cols}
\alias{.get_unique_cols}
\title{Get Unique Columns from Data Sources}
\usage{
.get_unique_cols(col_names)
}
\arguments{
\item{col_names}{A list of column names.}
}
\value{
A vector of unique column names.
}
\description{
Retrieves all unique columns from the data sources.
}
