% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.get_unique_levels}
\alias{.get_unique_levels}
\title{Get Unique Factor Levels}
\usage{
.get_unique_levels(factor_levels, level_conflicts)
}
\arguments{
\item{factor_levels}{A list of factor levels.}

\item{level_conflicts}{A list of variables with level conflicts.}
}
\value{
A list of unique factor levels.
}
\description{
Retrieves the unique factor levels for variables with conflicts.
}
