% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.get_var_classes}
\alias{.get_var_classes}
\title{Get Variable Classes from DataFrame}
\usage{
.get_var_classes(df.name, datasources)
}
\arguments{
\item{df.name}{Name of the input DataFrame.}

\item{datasources}{Data sources from which to aggregate data.}
}
\value{
A DataFrame containing the variable classes from each data source.
}
\description{
Retrieves the class of each variable in the specified DataFrame from different data sources.
}
