% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.identify_class_conflicts}
\alias{.identify_class_conflicts}
\title{Identify Class Conflicts}
\usage{
.identify_class_conflicts(classes)
}
\arguments{
\item{classes}{A DataFrame containing variable classes across data sources.}
}
\value{
A list of variables that have class conflicts.
}
\description{
Identifies conflicts in variable classes across different data sources.
}
