% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.identify_factor_vars}
\alias{.identify_factor_vars}
\title{Identify Factor Variables}
\usage{
.identify_factor_vars(var_classes)
}
\arguments{
\item{var_classes}{A DataFrame containing variable classes.}
}
\value{
A vector of factor variables.
}
\description{
Identifies which variables are factors in the DataFrame.
}
