% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.identify_level_conflicts}
\alias{.identify_level_conflicts}
\title{Identify Factor Level Conflicts}
\usage{
.identify_level_conflicts(factor_levels)
}
\arguments{
\item{factor_levels}{A list of factor levels.}
}
\value{
A list of variables with level conflicts.
}
\description{
Identifies conflicts in factor levels across different data sources.
}
