% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.make_levels_message}
\alias{.make_levels_message}
\title{Make Factor Level Conflict Message}
\usage{
.make_levels_message(level_conflicts)
}
\arguments{
\item{level_conflicts}{A list of variables with factor level conflicts.}
}
\value{
None. Prints the message to the console.
}
\description{
Creates a message to alert the user about factor level conflicts and prompt for action.
}
