% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.make_levels_recode_message}
\alias{.make_levels_recode_message}
\title{Make Levels Recode Message}
\usage{
.make_levels_recode_message(unique_levels)
}
\arguments{
\item{unique_levels}{A list of unique factor levels.}
}
\value{
A formatted string summarizing the level recoding.
}
\description{
Creates a message to alert the user about factor level recoding.
}
