% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_serverside_call}
\alias{.make_serverside_call}
\title{Create a Tidy Evaluation Call}
\usage{
.make_serverside_call(fun_name, df.name, tidy_select, other_args)
}
\arguments{
\item{fun_name}{The name of the function to be called (as a string), e.g., "select", "mutate".}

\item{df.name}{The name of the data frame (as a string) to which the function will be applied.}

\item{tidy_select}{Encoded tidyselect arguments}

\item{other_args}{A list of additional arguments to be passed to the function (optional).}
}
\value{
A call object that can be evaluated to perform the specified operation.
}
\description{
This function constructs a call object for a tidy evaluation function.
It allows for the dynamic creation of function calls in a tidyverse-compatible manner.
}
