% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.print_class_recode_message}
\alias{.print_class_recode_message}
\title{Print Class Recode Message}
\usage{
.print_class_recode_message(class_decisions, different_classes, newobj)
}
\arguments{
\item{class_decisions}{A vector of class decisions.}

\item{different_classes}{A list of variables with class conflicts.}

\item{newobj}{The name of the new DataFrame.}
}
\value{
None. Prints the message to the console.
}
\description{
Prints a message summarizing the class decisions that were made for variables with conflicts.
}
