% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.print_levels_recode_message}
\alias{.print_levels_recode_message}
\title{Print Factor Levels Recode Message}
\usage{
.print_levels_recode_message(unique_levels, newobj)
}
\arguments{
\item{unique_levels}{A list of unique factor levels.}

\item{newobj}{The name of the new DataFrame.}
}
\value{
None. Prints the message to the console.
}
\description{
Prints a message summarizing the factor level decisions that were made for variables with conflicts.
}
