% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.print_var_recode_message}
\alias{.print_var_recode_message}
\title{Print Variable Recode Message}
\usage{
.print_var_recode_message(added_cols, newobj)
}
\arguments{
\item{added_cols}{A list of added columns.}

\item{newobj}{The name of the new DataFrame.}
}
\value{
None. Prints the message to the console.
}
\description{
Prints a message summarizing the columns that were added to the DataFrame.
}
