% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.set_factor_levels}
\alias{.set_factor_levels}
\title{Set Factor Levels in DataFrame}
\usage{
.set_factor_levels(newobj, unique_levels, datasources)
}
\arguments{
\item{newobj}{The name of the new DataFrame.}

\item{unique_levels}{A list of unique factor levels.}

\item{datasources}{Data sources from which to aggregate data.}
}
\value{
None. Updates the DataFrame with the new factor levels.
}
\description{
Applies the unique factor levels to the DataFrame.
}
