% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.stop_if_cols_identical}
\alias{.stop_if_cols_identical}
\title{Stop If Columns Are Identical}
\usage{
.stop_if_cols_identical(col_names)
}
\arguments{
\item{col_names}{A list of column names from different data sources.}
}
\value{
None. Throws an error if columns are identical.
}
\description{
Checks if the columns in the data frames are identical and throws an error if they are.
}
