% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{.summarise_new_cols}
\alias{.summarise_new_cols}
\title{Summarize Newly Added Columns}
\usage{
.summarise_new_cols(newobj, datasources, col_names)
}
\arguments{
\item{newobj}{The name of the new DataFrame.}

\item{datasources}{Data sources from which to aggregate data.}

\item{col_names}{A list of column names.}
}
\value{
A list of added columns.
}
\description{
Summarizes the columns that were added to the DataFrame.
}
