% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.filter.R
\name{ds.filter}
\alias{ds.filter}
\title{Keep rows that match a condition}
\usage{
ds.filter(
  df.name = NULL,
  expr = NULL,
  .preserve = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df.name}{Character specifying a serverside data frame or tibble.}

\item{expr}{Expressions that return a logical value, and are defined in terms of the
variables in .data}

\item{.preserve}{Relevant when the .data input is grouped. If .preserve = FALSE (the default),
the grouping structure is recalculated based on the resulting data, otherwise the grouping is
kept as is.}

\item{newobj}{Character specifying name for new server-side data frame.}

\item{datasources}{DataSHIELD connections object.}
}
\value{
An object with the name specified by the \code{newobj} argument is written serverside.
}
\description{
DataSHIELD implentation of  \code{dplyr::filter}.
}
\examples{
\dontrun{
}
}
