% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.group_keys.R
\name{ds.group_keys}
\alias{ds.group_keys}
\title{Describe the groups within a grouped tibble or data frame}
\usage{
ds.group_keys(df.name = NULL, datasources = NULL)
}
\arguments{
\item{df.name}{Character specifying a server-side tibble.}

\item{datasources}{DataSHIELD connections object.}
}
\value{
A data frame describing the groups.
}
\description{
DataSHIELD implementation of \code{dplyr::group_keys}.
}
\examples{
\dontrun{
my_groups <- ds.group_keys("grouped_df")
}
}
