% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.select.R
\name{ds.select}
\alias{ds.select}
\title{Clientside dplyr select function}
\usage{
ds.select(.data = NULL, newobj = NULL, datasources = NULL, select_args = NULL)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{newobj}{Optional, character specifying name of new object}

\item{datasources}{datashield connections object}

\item{select_args}{Arguments normally passed to `...` in dplyr::select}
}
\value{
the object specified by the \code{newobj} argument or
of \code{ds.select} (or as default same name as input object) which is written to the serverside.
}
\description{
This function is similar to R function \code{select}.
}
\details{
Performs dplyr select
}
