% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.group_by.R
\name{ds.ungroup}
\alias{ds.ungroup}
\title{Remove grouping from a tibble or data frame}
\usage{
ds.ungroup(x = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{a tibble or data frame.}

\item{newobj}{Character specifying name for new server-side data frame.}

\item{datasources}{DataSHIELD connections object.}
}
\value{
No return value, called for its side effects. An ungrouped data frame or tibble is
created on the server.
}
\description{
DataSHIELD implentation of \code{dplyr::ungroup}.
}
\examples{
\dontrun{
ds.ungroup("grouped_df")
}
}
