% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{prompt_user_class_decision}
\alias{prompt_user_class_decision}
\title{Prompt User for Class Decision for a Single Variable}
\usage{
prompt_user_class_decision(var, all_servers, all_classes)
}
\arguments{
\item{var}{The variable name with a class conflict.}

\item{all_servers}{The names of all servers.}

\item{all_classes}{The classes of the variable across servers.}
}
\value{
A decision for the variable's class.
}
\description{
Prompts the user to resolve a class conflict for a single variable.
}
