% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{prompt_user_class_decision_all_vars}
\alias{prompt_user_class_decision_all_vars}
\title{Prompt User for Class Decision for All Variables}
\usage{
prompt_user_class_decision_all_vars(vars, all_servers, all_classes)
}
\arguments{
\item{vars}{A vector of variable names with class conflicts.}

\item{all_servers}{The names of all servers.}

\item{all_classes}{The classes of the variables across servers.}
}
\value{
A vector of decisions for each variable's class.
}
\description{
Prompts the user to resolve class conflicts for all variables.
}
