% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.group_by.R
\name{ds.group_by}
\alias{ds.group_by}
\title{Group by one or more variables}
\usage{
ds.group_by(
  df.name = NULL,
  expr,
  .add = FALSE,
  .drop = TRUE,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df.name}{Character specifying a serverside data frame or tibble.}

\item{expr}{Variables or computations to group by.}

\item{.add}{When FALSE, the default, group_by() will override existing groups. To add to the
existing groups, use .add = TRUE.}

\item{.drop}{Drop groups formed by factor levels that don't appear in the data? The default is
TRUE except when .data has been previously grouped with .drop = FALSE.}

\item{newobj}{Character specifying name for new server-side data frame.}

\item{datasources}{DataSHIELD connections object.}
}
\value{
An object with the name specified by the \code{newobj} argument is written serverside.
}
\description{
DataSHIELD implentation of \code{dplyr::group_by}.
}
\examples{
\dontrun{
 ds.filter(
}
}
