% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tidy_fill.R
\name{ds.tidy_fill}
\alias{ds.tidy_fill}
\title{Fill DataFrame with Missing Columns and Adjust Classes}
\usage{
ds.tidy_fill(df.name = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{df.name}{Name of the input DataFrame to fill.}

\item{newobj}{Name of the new DataFrame object created after filling.}

\item{datasources}{Data sources from which to aggregate data. Default is \code{NULL}.}
}
\value{
The filled DataFrame with added columns and adjusted classes or factor levels.
}
\description{
This function fills a given DataFrame by adding missing columns, ensuring consistent column classes, and adjusting factor levels where necessary.
It performs checks to detect class and factor level conflicts and prompts the user for decisions to resolve these conflicts.
}
