% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.group_by.R
\name{ds.ungroup}
\alias{ds.ungroup}
\title{Remove grouping}
\usage{
ds.ungroup(x = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{a tbl.}

\item{newobj}{Character specifying name for new server-side data frame.}

\item{datasources}{DataSHIELD connections object.}
}
\description{
DataSHIELD implentation of \code{dplyr::ungroup}.
}
\examples{
\dontrun{
 ds.ungroup("grouped_df")
 }
}
